+++ sources/objects/AssignmentUnitTests.cpp
@@ -1,4 +1,4 @@
-// AssignmentUnitTests.cpp by Ulrich Eisenecker, March 18, 2025
+// AssignmentUnitTests.cpp by Ulrich Eisenecker, May 20, 2026
 
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch_test_macros.hpp>
@@ -14,8 +14,8 @@
 #define COMPILE_FOR_TESTING
 
 #include "assignment.hpp"
-#include <string>
 #include <exception>
+#include <string>
 
 using namespace std;
 using namespace assignment;